Shellcheck ssh-keys feature

+++ optional-features/ssh-keys/feature.sh
@@ -4,7 +4,7 @@
 # Get absolute path to directory of script
 BASE_DIR="$( readlink -f "$(dirname "$0")" )"
 # Get absolute path to root of repo
-LAD_OS_DIR="$( echo $BASE_DIR | grep -o ".*/LadOS/" | sed 's/.$//')"
+LAD_OS_DIR="$( echo "$BASE_DIR" | grep -o ".*/LadOS/" | sed 's/.$//')"
 CONF_DIR="$LAD_OS_DIR/conf/ssh-keys"
 
 source "$LAD_OS_DIR/common/feature_header.sh"
@@ -16,16 +16,18 @@ feature_name="ssh-keys"
 feature_desc="Install existing ssh keys for your user and for root"
 
 provides=()
-new_files=("$HOME/.ssh/id_rsa" \
-    "$HOME/.ssh/config" \ 
-    "$HOME/.ssh/authorized_keys" \ 
-    "$HOME/.ssh/id_rsa.pub" \ 
+new_files=( \
+    "$HOME/.ssh/id_rsa" \
+    "$HOME/.ssh/config" \
+    "$HOME/.ssh/authorized_keys" \
+    "$HOME/.ssh/id_rsa.pub" \
     "$HOME/.ssh/known_hosts" \
-    "/root/.ssh/id_rsa" \ 
-    "/root/.ssh/config" \ 
-    "/root/.ssh/authorized_keys" \ 
-    "/root/.ssh/id_rsa.pub" \ 
-    "/root/.ssh/known_hosts")
+    "/root/.ssh/id_rsa" \
+    "/root/.ssh/config" \
+    "/root/.ssh/authorized_keys" \
+    "/root/.ssh/id_rsa.pub" \
+    "/root/.ssh/known_hosts" \
+)
 modified_files=()
 temp_files=()
 
@@ -54,37 +56,57 @@ function check_install() {
 }
 
 function install() {
-    if [[ -d "$CONF_DIR/user/.ssh" ]] &&
-        [[ -d "$CONF_DIR/root/.ssh" ]]; then
-        
+    if [[ -d "$CONF_DIR/user/.ssh" ]] && [[ -d "$CONF_DIR/root/.ssh" ]]; then
         qecho "Copying user ssh files to $HOME/.ssh..."
         mkdir -p "$HOME/.ssh"
         chmod 700 "$HOME/.ssh"
-        [[ -f "$USER_SSH_DIR/id_rsa" ]] &&
+
+        if [[ -f "$USER_SSH_DIR/id_rsa" ]]; then
             command install -Dm 600 "$USER_SSH_DIR/id_rsa" "$HOME/.ssh/id_rsa"
-        [[ -f "$USER_SSH_DIR/config" ]] &&
+        fi
+
+        if [[ -f "$USER_SSH_DIR/config" ]]; then
             command install -Dm 600 "$USER_SSH_DIR/config" "$HOME/.ssh/config"
-        [[ -f "$USER_SSH_DIR/authorized_keys" ]] &&
+        fi
+
+        if [[ -f "$USER_SSH_DIR/authorized_keys" ]]; then
             command install -Dm 600 "$USER_SSH_DIR/authorized_keys" "$HOME/.ssh/authorized_keys"
-        [[ -f "$USER_SSH_DIR/id_rsa.pub" ]] &&
+        fi
+
+        if [[ -f "$USER_SSH_DIR/id_rsa.pub" ]]; then
             command install -Dm 644 "$USER_SSH_DIR/id_rsa.pub" "$HOME/.ssh/id_rsa.pub"
-        [[ -f "$USER_SSH_DIR/known_hosts" ]] &&
+        fi
+
+        if [[ -f "$USER_SSH_DIR/known_hosts" ]]; then
             command install -Dm 644 "$USER_SSH_DIR/known_hosts" "$HOME/.ssh/known_hosts"
-        chown -R $USER "$HOME/.ssh"
+        fi
+
+        chown -R "$USER" "$HOME/.ssh"
 
         qecho "Copying root ssh files to /root/.ssh"
         sudo mkdir -p "/root/.ssh"
         sudo chmod 700 "/root/.ssh"
-        [[ -f "$ROOT_SSH_DIR/id_rsa" ]] &&
+
+        if [[ -f "$ROOT_SSH_DIR/id_rsa" ]]; then
             sudo install -Dm 600 "$ROOT_SSH_DIR/id_rsa" "/root/.ssh/id_rsa"
-        [[ -f "$ROOT_SSH_DIR/config" ]] &&
+        fi
+
+        if [[ -f "$ROOT_SSH_DIR/config" ]]; then
             sudo install -Dm 600 "$ROOT_SSH_DIR/config" "/root/.ssh/config"
-        [[ -f "$ROOT_SSH_DIR/authorized_keys" ]] &&
+        fi
+
+        if [[ -f "$ROOT_SSH_DIR/authorized_keys" ]]; then
             sudo install -Dm 600 "$ROOT_SSH_DIR/authorized_keys" "/root/.ssh/authorized_keys"
-        [[ -f "$ROOT_SSH_DIR/id_rsa.pub" ]] &&
+        fi
+
+        if [[ -f "$ROOT_SSH_DIR/id_rsa.pub" ]]; then
             sudo install -Dm 644 "$ROOT_SSH_DIR/id_rsa.pub" "/root/.ssh/id_rsa.pub"
-        [[ -f "$ROOT_SSH_DIR/known_hosts" ]] &&
+        fi
+
+        if [[ -f "$ROOT_SSH_DIR/known_hosts" ]]; then
             sudo install -Dm 644 "$ROOT_SSH_DIR/known_hosts" "/root/.ssh/known_hosts"
+        fi
+
         sudo chown -R root "/root/.ssh"
 
         qecho "Done copying ssh files"
@@ -95,7 +117,7 @@ function install() {
 }
 
 function uninstall() {
-    qecho "Removing ${new_files[@]}..."
+    qecho "Removing ${new_files[*]}..."
     rm -f "${new_files[@]}"
 }
